renamed Viz to Results and more standardizing of tab keys.
