+++ src/components/ResultsTab/Visualization/coastal_vulnerability/index.jsx
