+++ src/components/ResultsTab/Visualization/habitat_risk_assessment/Plot/Legend.js
