+++ src/components/ResultsTab/Visualization/habitat_risk_assessment/reducers/reducer_vectors_on_map.js
